+++ NEWS.md
@@ -1,4 +1,5 @@
 # spant 0.14.0
+* Added auto_phase function for zeroth order phase-correction of simple spectra.
 
 # spant 0.13.0
 * Various bug fixes for Siemens TWIX reader.